Safari-609.1.11.1/LayoutTests/accessibility/aria-labeled-with-hidden-node.html

<!DOCTYPE html>
<html>
<title id="title">TITLE</title>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>
<div id="content">
<button id="button" aria-labelledby="title">Go</button>
</div>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that if try to use a non-rendered node as your labelledby element, it won't crash and it will work.");
    
    if (window.accessibilityController) {
       var button = accessibilityController.accessibleElementById("button");
       shouldBeEqualToString("platformValueForW3CName(button)", "TITLE");

       document.getElementById("content").style.visibility = "hidden";
    }
</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
